+++ previous-versions/common/src/sdk/models/operations/getwebhook.ts
@@ -24,15 +24,15 @@ export class GetWebhookResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
-    /**
-     * OK
-     */
-    @SpeakeasyMetadata()
-    rule?: shared.Rule;
-
     @SpeakeasyMetadata()
     statusCode: number;
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    /**
+     * OK
+     */
+    @SpeakeasyMetadata()
+    webhook?: shared.Webhook;
 }