+++ previous-versions/common/RELEASES.md
@@ -651,3 +651,13 @@ Based on:
 - [typescript v0.36.1] previous-versions/common
 ### Releases
 - [NPM v0.36.1] https://www.npmjs.com/package/@codat/common/v/0.36.1 - previous-versions/common
+
+## 2023-09-05 00:03:40
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.0] previous-versions/common
+### Releases
+- [NPM v0.37.0] https://www.npmjs.com/package/@codat/common/v/0.37.0 - previous-versions/common